# docs/swagger.yaml
basePath: /api
definitions:
  SangXanh_pkg_dto.Address:
    properties:
      address_json:
        type: string
      is_default_address:
        type: boolean
      name:
        type: string
      phone:
        type: string
    type: object
...
  SangXanh_pkg_dto.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  SangXanh_pkg_dto.ResetPasswordRequest:
    properties:
      email:
        type: string
    type: object
  SangXanh_pkg_dto.UserInfo:
    properties:
      address:
        items:
          $ref: '#/definitions/SangXanh_pkg_dto.Address'
        type: array
      avatar:
        type: string
      basic_address:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        $ref: '#/definitions/SangXanh_pkg_enum.Status'
      username:
        type: string
    type: object
  SangXanh_pkg_dto.UserRegisterRequest:
    properties:
      avatar:
        type: string
      basic_address:
        type: string
      email:
        type: string
      full_name:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  SangXanh_pkg_dto.UserUpdateAddressRequest:
    properties:
      address:
        items:
          $ref: '#/definitions/SangXanh_pkg_dto.Address'
        type: array
      id:
        type: string
    type: object
  SangXanh_pkg_dto.UserUpdateData:
    properties:
      id:
        type: string
      status:
        $ref: '#/definitions/SangXanh_pkg_enum.Status'
    type: object
  SangXanh_pkg_dto.UserUpdateRequest:
    properties:
      avatar:
        type: string
      basic_address:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      phone:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
  SangXanh_pkg_enum.Status:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - Active
    - Inactive
host: localhost:8080
info:
  contact:
    email: support@sangxanh.com
    name: API Support
    url: http://www.sangxanh.com/support
  description: This is the SangXanh API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SangXanh API
  version: "1.0"
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users with optional filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by role
        in: query
        name: role
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of users
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: List all users
      tags:
      - Users
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.UserInfo'
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      summary: Get user by ID
      tags:
      - Users
  /user/address:
    put:
      consumes:
      - application/json
      description: Update the authenticated user's addresses
      parameters:
      - description: Address data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.UserUpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user with new addresses
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.UserInfo'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user address
      tags:
      - Users
  /user/change-password:
    put:
      consumes:
      - application/json
      description: Change the authenticated user's password
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: Password changed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized or invalid old password
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - Users
  /user/forgot-password:
    put:
      consumes:
      - application/json
      description: Send password reset instructions to user's email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reset user password
      tags:
      - Users
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.UserInfo'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Users
  /user/send-magic-link:
    put:
      consumes:
      - application/json
      description: Send a magic link to user's email for passwordless authentication
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Magic link sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      summary: Send magic link for passwordless login
      tags:
      - Users
  /user/update:
    put:
      consumes:
      - application/json
      description: Update authenticated user's profile information
      parameters:
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.UserInfo'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "429":
          description: Too many requests
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user information
      tags:
      - Users
  /user/update-status:
    put:
      consumes:
      - application/json
      description: Update a user's status (admin only)
      parameters:
      - description: User status data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SangXanh_pkg_dto.UserUpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: Updated user
          schema:
            $ref: '#/definitions/SangXanh_pkg_dto.UserInfo'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - admin only
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"